requirements.txt
Flask==1.1.2
Jinja2==2.11.2
Werkzeug==1.0.1
joblib==0.16.0
numpy==1.19.2
tensorflow-cpu==2.4.1
pip==20.2.3
Pillow==7.1.2
requests==2.24.0
scipy==1.5.2
setuptools==41.2.0
threadpoolctl==2.1.0
urllib3==1.25.10
wheel==0.35
gunicorn==20.0.4
opencv-python-headless==4.5.1.48
pandas==1.1.3